Br trung merge br_Thong

+++ QLTHUCHI/src/model/objects/LogO.java
@@ -12,7 +12,7 @@ import java.util.Date;
 public class LogO {
     private int ID_Log;
     private int ID_Type;
-    private int Price;
+    private double Price;
     private String Note;
     private String DateString;
     private int User_ID;
@@ -25,7 +25,7 @@ public class LogO {
         this.ID_Log = -1;
     }
 
-    public LogO(int ID_Log, int ID_Type, int Price, String Note, String DateString) {
+    public LogO(int ID_Log, int ID_Type, double Price, String Note, String DateString) {
         this.ID_Log = ID_Log;
         this.ID_Type = ID_Type;
         this.Price = Price;
@@ -46,11 +46,11 @@ public class LogO {
         this.ID_Type = ID_Type;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return Price;
     }
 
-    public void setPrice(int Price) {
+    public void setPrice(double Price) {
         this.Price = Price;
     }
